:recycle: refactor(web): repoint billing routes and consumers to the commercial slot
Finish moving billing behind the commercial overlay slot: every billing API
route and every entitlement/pricing consumer now imports from
@/lib/extensions/billing-* instead of @/lib/billing/* and @/components/billing/*.
A build without the commercial overlay resolves them all the same way.

The ai-entitlement route drops its inline auth check, envelope unwrapping and
error handling — the same work proxyBilling already does — and just forwards
through the slot like the other routes.
