+++ apps/web/src/app/api/billing/plans/route.ts
@@ -1,4 +1,4 @@
-import { proxyBilling } from '@/lib/billing/proxy';
+import { proxyBilling } from '@/lib/extensions/billing-proxy';
 
 /** Enabled plans (credit packs + subscriptions) for the pricing page. */
 export function GET() {
