+++ apps/web/src/app/api/billing/subscriptions/route.ts
@@ -1,7 +1,12 @@
 import { NextRequest } from 'next/server';
-import { proxyBilling } from '@/lib/billing/proxy';
+import { proxyBilling } from '@/lib/extensions/billing-proxy';
 
-/** Create a subscription ({ planId, channel:'paypal' }) → returns approve URL. */
+/**
+ * Create a subscription → returns the payment provider's approval URL.
+ *
+ * The body is forwarded untouched; which provider and plan shape it carries is
+ * the commercial build's business, not this one's.
+ */
 export async function POST(req: NextRequest) {
   const body = await req.text();
   return proxyBilling('/api/billing/subscriptions', { method: 'POST', body });
